Rename viridian to cylon
