+++ .github/workflows/build.yaml
@@ -1,4 +1,4 @@
-name: Build viridian images
+name: Build cylon images
 
 on:
   push:
@@ -68,7 +68,7 @@ jobs:
     - name: Build Container
       env:
         SINGULARITY_RECIPE: Singularity.def
-        OUTPUT_CONTAINER: viridian_${{env.RELEASE_VERSION}}.img
+        OUTPUT_CONTAINER: cylon_${{env.RELEASE_VERSION}}.img
       run: |
         ls
         if [ -f "${SINGULARITY_RECIPE}" ]; then
@@ -83,6 +83,6 @@ jobs:
       if: startsWith(github.event.ref, 'refs/tags/v')
       uses: softprops/action-gh-release@v1
       with:
-        files: viridian_*.img
+        files: cylon_*.img
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
