+++ cylon/__init__.py
@@ -1,7 +1,7 @@
 from pkg_resources import get_distribution
 
 try:
-    __version__ = get_distribution("viridian").version
+    __version__ = get_distribution("cylon").version
 except:
     __version__ = "local"
 
@@ -15,4 +15,4 @@ __all__ = [
     "utils",
 ]
 
-from viridian import *
+from cylon import *
