+++ cylon/tasks/__init__.py
@@ -2,4 +2,4 @@ __all__ = [
     "assemble",
 ]
 
-from viridian.tasks import *
+from cylon.tasks import *
