+++ tests/utils_test.py
@@ -3,7 +3,7 @@ import pytest
 
 import pyfastaq
 
-from viridian import utils
+from cylon import utils
 
 this_dir = os.path.dirname(os.path.abspath(__file__))
 data_dir = os.path.join(this_dir, "data", "utils")
