+++ tox.ini
@@ -12,4 +12,4 @@ deps =
     pytest-cov
     -r requirements.txt
 commands =
-    pytest --cov=viridian --cov-append --cov-report=term-missing
+    pytest --cov=cylon --cov-append --cov-report=term-missing
